+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.5.14
 
+* `Coroutines`
+    * `subscribeSafelyWithoutExceptions` got new parameter `onException` by analogue with `safelyWithoutExceptions`
+    * New extensions `Flow#subscribeAsync` and subsequent analogs of `subscribe` with opportunity to set up custom marker
+
 ## 0.5.13
 
 * `Common`: